requirements.txt
streamlit==1.56.0
langchain==1.2.15
langchain-core==1.3.1
langchain-ollama==1.1.0
langchain-google-genai==4.2.2
langchain-openai==1.2.0
langchain-anthropic>=0.3.0
python-dotenv==1.2.2
